// app/turkish-teachers/turkish-teacher-types.ts
export type TurkishTeacher = {
    id: string
    name: string
    surname: string
    whatsapp_number: string
    gender: "male" | "female"
    age: number
    city: string
    address: string
    experience_years: number
    notes: string | null
    created_at: string
    updated_at: string
}